Use ByteBuffer instead of byte[].
AbstractByteBufferAllocator gives the libuv library a byte array,
it is then filled with data and the amount of written data is then
returned to the allocator back. The Allocator doesn't know the exact
amount bytes to allocate before libuv writes it into the buffer, libuv
just suggests a size. After the buffer has been written, the Allocator
is notified of the amount of bytes written into it.