Minimize API.
This solves the problem of namespace overusage (I have already
a function called Read and how do I name the event? ReadEvent?).

The other thing is that an event variable might indicate to the
user that he can add multiple callbacks and alter the returned
data in one of the callbacks which might lead to unwanted
behaviour in other callbacks (because they all share the same
data).