Expose uv_*_open functionality.
Will let one use an already existing OS socket.
Now I had to make connect an instance method in order
to make Open useable at all.
It turned out that this resulted in a significant cut
down of overload methods(no multiplexing of the Loop argument
required, only in the constructor).
This is results in less methods, but the user has now to care
about the order of methods he is invoking.