Save current loop when using Run and return it with Current.
This way even if we run the loop normally (not in the
Loop.Run(Func<Task> overload) we ensure that Loop.Current
will return the currently running loop.

However, the context is still useful if we want to write
counter parts of the evented APIs which block but still spin the
loop.