Rename Async and FAsync to QueueUSerWorkItemAsync.
Since that is what it basically is. A wrapped version
of QueueUserWorkItemAsync.

The names Async and FAsync are not really clear cut in
what they do. I tried to name them as a counter part
to Sync, which Synchronizes something from an external
thread into the event loop (it runs a delegate in the
event loop).