Update travis build.
It has now the csharp language, so the configuration is smaller.
It also runs now with sudo: false, which makes travis
use their new container infrastructure and is far more faster.

Also the nice level in the travis container environment is always 0,
so I removed it from the tests.

Thanks @akoeplinger for helping!