Fix Task.Status being set to Running on windows/net.
It is impossible to set the Task.Status via some API. We need to set it
using
Reflection on different platforms differently. On mono the field is
called
status and can be set directly with TaskStatus.Running, on Windows the
field
is called m_stateFlags and a bit flag has to be added.