Fix performance issues.
Do not feed dynamically created delegates to pinvoke
methods, the runtime will have to create a stub for
each of them.

It is better to create a static method and then use
it, since only one callback stub will be created

Note that the process code was badly written, we
have overwritten the process DataPointer again.
We didn't need to do so and this error only
appeared after making CloseCallback a static method.